accountability_project/requirements.txt
setuptools==78.1.0
cffi==1.17.1
pycparser==2.22
Django==5.2
django-cors-headers==4.7.0
django-simple-history==3.8.0
djangorestframework==3.16.0
psycopg==3.2.6
django-model-utils==4.5.1
django.filter==25.1
firebase-admin==7.2.0

# libraries for API documentation (just development)
pyyaml==6.0.2
drf-spectacular==0.28.0
# other development libs
autopep8==2.3.2
